# Use an iterator-style API to consume shaped text in line wrapper