@whereswaldon 2023-05-04
Please review the iterator portion of this changeset, but hold off on reviewing the allocation optimization. I couldn't sleep tonight because I thought of a much better way to structure that, and it's super promising. I need a little more time to get it right though. I will update the PR and benchmarks once I'm happy with it.

At a high level, the current optimization exposes too many details to the caller (and asks them to understand them) *and* isn't able to adaptively resize the pre-allocated buffers depending upon workload. I'm internalizing everything but the existence of a buffer type and making it grow adaptively when the wrapping workload consumes all resources. We can tune the growth strategy without changing the public API, which seems like a much better place to be. This strategy is able to almost completely eliminate allocations when wrapping multiple large texts consecutively, as the buffer growth for wrapping the first one will generally be sufficient to handle the second.