@whereswaldon 2023-05-16
Okay, I've updated this branch with my latest work. I decided that separating out #69 didn't really add as much value as I hoped, so those changes are incorporated here as well. I hope that the result is much easier to understand. I've updated the PR description with latest benchmark results. On the whole, this is a perf win and a feature win, but we do lose some performance in places.