@whereswaldon 2023-05-19
> A couple of comments on naming.
> It all looks sensible, but I cannot figure out how the wrap buffer is worked with - there is no public API for accessing, copying, or resetting its data.
> Are we just to use the objects as a black box and manage as many of them as we need?

Precisely. They are opaque to users of the package, but exposed in order to allow callers to control whether they invalidate the last lines returned from the wrapper by reusing the buffer. If we don't expose this concept, callers cannot wrap a new paragraph until they are done with the previous (is that an okay constraint? I thought not). If we don't allow this kind of buffer reuse then we allocate quite a lot in wrapping each time.