@whereswaldon 2023-05-19
> Is there any sensible default we can use for buffering (without hitting the "stop until buffer copied" issue you raised above)?
> It feels like having an API that is less performant by default is a step backwards.
> I don't disagree that these are all desirable improvements - but I wonder if we can retain the default performance from before in some way?

I'm not sure what "stop until buffer copied" issue you mean. I just reread the conversation above and I didn't see anything that seemed to match.

We can retain more performance by complicating the iterator API, but I opted to keep it simple in this iteration. Allowing the wrapper to ask "is there only one run?" allows us to keep performance on par.