@benoitkugler 2023-05-20
Regarding the buffer, could we merge it with the `LineWrapper` type, and document that the lines returned by Wrap are owned by the line wrapper ?
User who need multiple wrapping would create multiple `LineWrapper`. And we could completely hide the scratch type from the API, and have buffered, performant behavior by default.