@whereswaldon 2023-05-22
I've internalized the existence of `WrapBuffer` as @benoitkugler suggested and tried to address all of @andydotxyz's recommendations regarding naming and style. Performance is essentially unchanged, though there is no longer any need to benchmark unbuffered use of the API because the buffering is internal and automatic.