@benoitkugler 2023-05-22
Late remark : since `wrapBuffer` is now hidden, we could probably remove the pointer indirection (in the `LineWrapper.scratch` field), and also simplify `ensureScratch` that could just be replaced by `scratch.reset`.

@whereswaldon I can commit such changes if you want.