@whereswaldon 2023-05-22
I've implemented that change, with some extra logic to make sure that the first time we reset the scratch buffer it allocates *some* space so that the first use of the line wrapper isn't extra-slow for no reason.