@andydotxyz 2023-05-29
> I'm not sure what "stop until buffer copied" issue you mean. I just reread the conversation above and I didn't see anything that seemed to match.

Sorry I was referring to the same concept that you better described as "callers cannot wrap a new paragraph until they are done with the previous". I think I misunderstood what the default behaviour is so it should be fine.